Refactored virtual key handling logic out of Terminal.
